test: split state tests by type to 3
